apps/server/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "SohamRatnaparkhi",email = "soham.ratnaparkhi@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard] (>=0.115.8,<0.116.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "google-generativeai (>=0.8.4,<0.9.0)",
    "google-genai (>=1.3.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
